Refactor initiated
Now all old files are in the '__OLD' directoy.
Files related to the RFEP (Random Forest Evolutionary Prunning) are now in the 'rfep' module and are available to import in simplified use. Right now classes for model and data persist, for evaluation configuration and for the genetic algorithm are ready for use.
More code is on the way for result analisis in the new version.